Separate v2 and v4 filter factory
