+++ packages/odata-v4/test/test-util/index.ts
@@ -1,0 +1,2 @@
+export * from './filter-factory';
+export * from './test-util';
