+++ test-resources/test/test-util/index.ts
@@ -1,9 +1,9 @@
 export * from './request-mocker';
+export * from './custom-de-serializers';
 export * from './batch-test-util';
 export * from './destination-service-mocks';
 export * from './environment-mocks';
 export * from './example-destination-service-responses';
-export * from './filter-factory';
 export * from './keys';
 export * from './mocked-access-tokens';
 export * from './test-data';
